feat(*): Docs, additional API features (#31)

+++ keel-front50/src/main/kotlin/com/netflix/spinnaker/keel/front50/Front50IntentRepository.kt
@@ -48,7 +48,14 @@ class Front50IntentRepository
 
   override fun getIntent(id: String) = front50Service.getIntent(id)
 
-  override fun deleteIntent(id: String) {
-    front50Service.deleteIntent(id)
+  override fun deleteIntent(id: String, preserveHistory: Boolean) {
+    if (preserveHistory) {
+      getIntent(id).also {
+        it.status = IntentStatus.DELETED
+        upsertIntent(it)
+      }
+    } else {
+      front50Service.deleteIntent(id)
+    }
   }
 }
